Split aggregate budget handling from loop stream
